Add log files
Include log.hpp in common.hpp
