+++ src/AvTranscoder/common.hpp
@@ -10,15 +10,10 @@ extern "C" {
 	#define UINT64_C(c) (c ## ULL)
 #endif
 #include <libavcodec/version.h>
-#include <libavutil/error.h>
 #include <libavutil/rational.h>
-#include <libavutil/log.h>
 }
 
 #include <string>
-#include <cstring>
-#include <sstream>
-#include <fstream>
 
 #ifdef SWIG
  #define AvExport
@@ -50,6 +45,8 @@ extern "C" {
 #endif
 #endif
 
+#include <AvTranscoder/log.hpp>
+
 namespace avtranscoder
 {
 
@@ -62,48 +59,6 @@ void AvExport preloadCodecsAndFormats();
 
 /// Get the string description corresponding to the error code provided by ffmpeg/libav
 std::string AvExport getDescriptionFromErrorCode( const int code );
-
-#define LOG_DEBUG( ... ) { std::stringstream os; os << __VA_ARGS__; Logger::log( AV_LOG_DEBUG, os.str() ); }
-#define LOG_INFO( ... ) { std::stringstream os; os << __VA_ARGS__; Logger::log( AV_LOG_INFO, os.str() ); }
-#define LOG_WARN( ... ) { std::stringstream os; os << __VA_ARGS__; Logger::log( AV_LOG_WARNING, os.str() ); }
-#define LOG_ERROR( ... ) { std::stringstream os; os << __VA_ARGS__; Logger::log( AV_LOG_ERROR, os.str() ); }
-
-/// Logger class which contains static functions to use ffmpeg/libav log system
-class AvExport Logger
-{
-public:
-	/**
-	 * @brief Set the log level of ffmpeg/libav.
-	 * @param level: refer to define AV_LOG_xxx (from AV_LOG_QUIET to AV_LOG_DEBUG)
-	 * @see SWIG interface avLogLevel.i
-	 */
-	static void setLogLevel( const int level );
-
-	/**
-	 * @brief Log with the ffmpeg/libav log system
-	 * @note use define LOG_* to log at DEBUG/INFO/WARN/ERROR level
-	 * @param msg: the message will be prefixed by '[avTranscoder - <level>]'
-	 * @param msg: the message will be suffixed by '\n'
-	 */
-	static void log( const int level, const std::string& msg );
-
-	/**
-	 * @brief Log ffmpeg/libav and avtranscoder informations in a text file.
-	 * @note Default log filename is avtranscoder.log
-	 * @see getLogFileName
-	 * @see setLogFileName
-	 */
-	static void logInFile();
-
-	///@{
-	/// @warning Need to set the expected log filename before calling logInFile
-	static std::string& getLogFileName() { return _logFileName; }
-	static void setLogFileName( const std::string& newLogFileName ) { _logFileName = newLogFileName; }
-	///@}
-
-private:
-	static std::string _logFileName;  ///< Name of the log file
-};
 
 }
